Fix high memory consumption by creating new Document for every record
The default allocator of the rapidjson::Document does not free its memory when
the document is set to null, but when the object is destroyed. This can lead to
several GiB of memory usage when parsing large WARC files.
This is fixed by creating a new Document for every parsed WARC record.
Another option would be to use the CrtAllocator instead of the
MemoryPoolAllocator which frees its memory directly when free is called.